Adds redirect to home page in case of invalid ID in note editing and deletion functions

+++ app/Http/Controllers/MainController.php
@@ -51,6 +51,9 @@ class MainController extends Controller
     public function editNote($id)
     {
         $id = Operations::decryptId($id);
+        if ($id == null) {
+            return redirect()->route('home');
+        }
         $note = Note::find($id);
         return view('notes.edit_note', [
             'note' => $note,
@@ -79,6 +82,9 @@ class MainController extends Controller
         }
 
         $id = Operations::decryptId($request->note_id);
+        if ($id == null) {
+            return redirect()->route('home');
+        }
         $note = Note::find($id);
         $note->title = $request->text_title;
         $note->text = $request->text_note;
@@ -90,6 +96,9 @@ class MainController extends Controller
     public function deleteNote($id)
     {
         $id = Operations::decryptId($id);
+        if ($id == null) {
+            return redirect()->route('home');
+        }
         $note = Note::find($id);
         return view('notes.delete_note', [
             'note' => $note,
@@ -99,6 +108,9 @@ class MainController extends Controller
     public function deleteNoteConfirm($id)
     {
         $id = Operations::decryptId($id);
+        if ($id == null) {
+            return redirect()->route('home');
+        }
         $note = Note::find($id);
         $note->delete();
